Experimenting with making ordinals cumulative
It is rather messy :-(
Cumulative inductives work for Aczel trees but fail when quotienting; this might be an incompleteness in the cumulativity analysis.
Coq's attempt at infering universes often adds too many universes.
It tries to be as general as possible.
This adds too many unnessary universes.
Coq's universe polymorphism with monomorphic univrses and their constraints makes it even worse.